feat: Created a config key to define custom containers and analytics url
This features wants to provide a way to custom the entrypoint to static
containers and analytics.
In a high traffic machine, analytics will use too much resource and it's
natural to use it behind a load balancer, but provide the static files on
the same serve may take a long time to response impacting on sites
performance.