Set idSite on tracker already when creating the tracker
Since Matomo 3.14.0 the cookies are no longer when setting idSite allowing us to configure the idSite early.

As a result, when the `TrackerAdded` event is triggered, then all the listeners will receive the tracker fully configured with idSite and tracker URL instead of only the URL. In particular this fixes eg Heatmaps & Session Recording when there are 2 trackers used in one container. Then currently, the second tracker was not requesting `configs.php` and was not checking for heatmaps or session recordings because at the time `TrackerAdded` was triggered, there was not yet an idSite configured.